Add Github Action to autogenerate requirements
Pipenv is a great tool for handling virtual environments and dependencies, however, the current python standard only considers pip as the default tooling, this PR adds an action that will generate the file automatically without affecting the dev process.

If needed, we can change the auto generation to run only on the main branch.